+++ src/main/java/de/samply/security/SecurityConfiguration.java
@@ -1,6 +1,7 @@
 package de.samply.security;
 
 import de.samply.app.ProjectManagerConst;
+import de.samply.user.roles.MethodRoles;
 import de.samply.user.roles.RolesExtractor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -11,6 +12,7 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configurers.AuthorizeHttpRequestsConfigurer;
 import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserService;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 import java.util.Map;
 
@@ -40,8 +42,10 @@ public class SecurityConfiguration {
     }
 
     private AuthorizeHttpRequestsConfigurer<HttpSecurity>.AuthorizationManagerRequestMatcherRegistry addAuthorityMapping(AuthorizeHttpRequestsConfigurer<HttpSecurity>.AuthorizationManagerRequestMatcherRegistry authorization) {
-        Map<String, String[]> pathRolesMap = RolesExtractor.extractPathRolesMap();
-        pathRolesMap.keySet().forEach(path -> authorization.requestMatchers(path).hasAnyAuthority(pathRolesMap.get(path)));
+        Map<String, MethodRoles> pathRolesMap = RolesExtractor.extractPathRolesMap();
+        pathRolesMap.keySet().forEach(path -> authorization
+                .requestMatchers(new AntPathRequestMatcher(path, pathRolesMap.get(path).httpMethod()))
+                .hasAnyAuthority(pathRolesMap.get(path).roles()));
         return authorization.anyRequest().authenticated();
     }