+++ CHANGELOG.md
@@ -31,3 +31,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Document Service
 - Upload and Download documents
 - Project Type
+- HTTP Method for every endpoint in Security Configuration